requirements.txt
django==5.0.6
django-bootstrap5==24.2
numpy==1.26.4
matplotlib==3.9.0
django-autocomplete-light==3.11.0
requests==2.32.3
tldextract==5.1.2
# Use fork of django-background-tasks which doesn't include "import imp" in tasks.py as that is removed in Py 3.12
django4-background-tasks @ git+https://github.com/LoyalMed/django4-background-tasks.git@3a844eb
psutil==5.9.8
impacket==0.11.0
pyasn1==0.6.0
fontawesomefree==6.5.1
django-datatables-view==1.20.0
neo4j==5.20.0
dateparser==1.2.0
django-taggit==5.0.1
XlsxWriter==3.2.0
jsonschema==4.22.0
json2table==1.1.5
django-reversion==5.0.12
django-taggit-bulk==0.0.4b1
django-referrer-policy==1.0
Pillow==10.2.0
duckdb==0.10.1
django_csp==3.8
django-permissions-policy==4.19.0
ansi2html==1.9.1
django-plugins-bihealth==0.5.2
